requirements.txt
piexif
python-docx
PyPDF2